Splitting ResultReporter header.
